Fix for stalled connections to broken Galera cluster nodes
There is a bug in MariaDB / Galera where sometimes a node will get stuck. It'll accept new TCP connections, but will never send any data from the client to the server. PHP's CONNECT TIMEOUT never happens, even though it is still in the CONNECT API call, as a valid TCP connection does indeed happen. Using a short READ TIMEOUT allows for PUDL to detect this particular connection and gracefully move on to the next potential node in the cluster.