Compatibility shim for $db->upsert()
NOTE: replace() is NOT always desired, this COULD in theory cause issues with foreign keys. Sqlite now supports a more complex "ON CONFLICT" clause, but this will require updating the underlaying PUDL API to support properly Sqlite requiring naming the unique key columns, so this must be dynamically queried and inserted into the generated query.