Version bump to 3.0.0 and RELEASE!!
This has been a long time coming, honestly.

The ONLY reason why PUDL has not seen a new release in years is that
historically, the version number was tied to the release schedule
of the Altaform project.

With version 3.0.0 and going forward, PUDL is now fully independent.

This means that future releases should happen quickly once again!