Benchmarking now returns start/stop instead of duration
The benchmarking system previously was used for just getting the elapsed time of a particular query. Instead, we are now returning the beginning and end time for a query, and allowing the application to calculate that information as it sees fit. This enables additional metrics, such as measuring PUDL library overhead or "time to first query start". These should be left up to the application to decide what is valuable to track, not left up to this library.